x86/cpu/topology: Move registration out of APIC code
commit c0a66c2847908e41c771ca2355fba935a82a9f62 upstream.

The APIC/CPU registration sits in the middle of the APIC code. In fact this
is a topology evaluation function and has nothing to do with the inner
workings of the local APIC.

Move it out into a file which reflects what this is about.

Intel-SIG: commit c0a66c284790 x86/cpu/topology: Move registration out of APIC code.
x86/apic: Rework APIC registration

Tested-by: Michael Kelley <mhklinux@outlook.com>
Tested-by: Sohil Mehta <sohil.mehta@intel.com>
Link: https://lore.kernel.org/r/20240213210251.543948812@linutronix.de
[ Quanxian Wang: amend commit log ]
